Removes concrete product models from longclaw, allowing customisation. For #76 and #47
