+++ longclaw/project_template/products/models.py
@@ -1,19 +1,41 @@
 from django.db import models
+from django_extensions.db.fields import AutoSlugField
 from modelcluster.fields import ParentalKey
+from wagtail.wagtailcore.models import Page
 from wagtail.wagtailcore.fields import RichTextField
-from longclaw.longclawproducts.models import ProductVariantBase, Product
+from wagtail.wagtailadmin.edit_handlers import FieldPanel
+from longclaw.longclawproducts.models import ProductVariantBase, ProductBase
+
+class ProductIndex(Page):
+    '''Index page for all products
+    '''
+    subpage_types = ('products.Product', 'products.ProductIndex')
+
+
+class Product(ProductBase):
+    parent_page_types = ['products.ProductIndex']
+    description = RichTextField()
+    content_panels = ProductBase.content_panels + [
+        FieldPanel('description'),
+    ]
+
+
+    @property
+    def first_image(self):
+        return self.images.first()
 
 class ProductVariant(ProductVariantBase):
 
-    # Set 'Product' as the parent of product variants.
-    # You can support your own 'Product' (and 'ProductIndex') pages
-    # by change the referenced model below to your own, or
-    # do away with the 'Product' concept entirely - e.g. if you only
+    # You *could* do away with the 'Product' concept entirely - e.g. if you only
     # want to support 1 'variant' per 'product'.
     product = ParentalKey(Product, related_name='variants')
 
+    slug = AutoSlugField(
+        separator='',
+        populate_from=('product', 'ref'),
+        )
+
     # Enter your custom product variant fields here
     # e.g. colour, size, stock and so on.
-    # Remember, ProductVariantBase provides 'price', 'ref', 'slug' fields
-    # and the parental key to the Product model.
+    # Remember, ProductVariantBase provides 'price', 'ref' and 'stock' fields
     description = RichTextField()
